restart pmcd manually based on task status
Do not use a handler for pmcd restart since it needs to be restarted
immediately if there are changes.  Instead, use `register` to collect
the changed status from all relevant tasks, and restart pmcd if needed,
and only if needed.  Otherwise, the role will only ensure that pmcd is
enabled and started.