Rename the embedded roles to remove 'performancecopilot' naming
We currently use embedded role names with the additional prefix
of 'performancecopilot_metrics_', used in the related collection.
This is to obfuscate the names of these roles in case someone was
to invoke them directly within the linux-system-roles.metrics
namespace.  Perhaps also someday we'd want those shorter, clearer
names for other embedded roles.

With recent changes, each role is now passed the metrics_provider
value (currently 'pcp') allowing per-role customisation for other
providers.  There are early plans to allow 'sysstat' to become an
alternate provider with some roles.  The long names are also very
unwieldy by design, which makes work in the existing metrics role
to be more painful than need be - for no reason anymore.

This commit renames the embedded roles back to the shortened form
to reflect the above state of affairs.