Reorder tasks such that Redis started before PCP services
Current versions of the pmproxy daemon in PCP attempt to
connect to Redis immediately on startup and disable some
functionality if that fails.

This should be fixed in PCP in due course but there's no
harm in starting Redis service first in the metrics role
(if metric_query is requested) to avoid the problem.