fix: Use correct syntax for keyserver port and test check
Cause: The role was using incorrect syntax for specifying the
keyserver port for the firewall.

Consequence: The role would give an error when configuring
the keyserver port in the firewall.

Fix: Use the correct syntax in the role and in the test
used to check the firewall.

Result: The role can configure the firewall port for the
keyserver correctly.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>