feat: Support this role in container builds
Feature: Support running the cockpit role during container builds.

Reason: This is particularly useful for building bootc derivative OSes.

Result: The role now works during container builds. The bootc container
scenarios run in CI, which ensures that the role works in buildah build
environment. This allows us to officially support this role for image
mode builds.

Detect if the system is booted (with systemd), and skip all runtime
operations and checks if not. Also use `firewall-offline-cmd` which
works also in non-booted environments -- we are only/primarily
interested in the persistent firewall config anyway.

Make tests_verify_fullstack.yml a full bootc end-to-end test, as that
covers the most code paths.

https://issues.redhat.com/browse/RHEL-98911